examples/embedded-devices/entry/mqtt-quic-python/requirements.txt
# TESAIoT MQTT over QUIC - Python Requirements
# =============================================
#
# Install: pip install -r requirements.txt
# Or use: ./install_deps.sh (recommended)
#

# Core QUIC library (cross-platform: macOS, Linux, Windows, RPi)
aioquic>=0.9.25

# Environment file support
python-dotenv>=1.0.0

# Optional: Standard MQTT for fallback (TCP+TLS)
paho-mqtt>=1.6.1

# Optional: For advanced SSL certificate handling
pyOpenSSL>=23.0.0
